# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cybersecurity Platform API
  version: 1.0.0
  description: API contract for public intake, secure client portal, and role-protected administration.
servers:
  - url: /api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: portal_session
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
...
paths:
...
  /public/newsletter:
    post:
      summary: Request a newsletter subscription
      responses:
        { '201': { description: Pending confirmation }, '422': { description: Validation error } }
...
  /public/career:
    post:
      summary: Apply to a currently open career opportunity
      responses:
        {
          '201': { description: Application persisted },
          '404': { description: Role not open },
          '422': { description: Validation error },
        }
...
  /portal/summary:
    get:
      security: [{ sessionCookie: [] }]
      summary: Organisation-scoped client overview
      responses:
        {
          '200': { description: Overview },
          '401': { description: Not authenticated },
          '403': { description: No organisation access },
        }
  /portal/tickets:
    get:
      security: [{ sessionCookie: [] }]
      summary: List support tickets within the authenticated organisation
      responses: { '200': { description: Organisation ticket list } }
    post:
      security: [{ sessionCookie: [], csrfHeader: [] }]
      summary: Create an idempotent support ticket within the authenticated organisation
      responses:
        {
          '201': { description: Ticket persisted },
          '401': { description: Not authenticated },
          '403': { description: No organisation access or invalid CSRF token },
          '422': { description: Validation error },
        }
  /portal/reports/{id}:
    get:
      security: [{ sessionCookie: [] }]
      summary: Read report metadata within the authenticated organisation
      parameters: [{ in: path, name: id, required: true, schema: { type: string } }]
      responses:
        {
          '200': { description: Report metadata },
          '404': { description: Not found or outside scope },
        }
...